refactor(memory): extract sub-components into separate files
